+++ stac_fastapi/extensions/stac_fastapi/extensions/core/collection_search/request.py
@@ -40,21 +40,21 @@ class BaseCollectionSearchPostRequest(BaseModel):
     bbox: Optional[BBox] = Field(
         default=None,
         description="Only return items intersecting this bounding box. Mutually exclusive with **intersects**.",  # noqa: E501
-        json_schema_extra={
-            "example": [-175.05, -85.05, 175.05, 85.05],
+        openapi_examples={
+            "user-provided": {"value": None},
+            "Montreal": {"value": "-73.896103,45.364690,-73.413734,45.674283"},
         },
     )
     datetime: Optional[str] = Field(
         default=None,
         description="""Only return items that have a temporal property that intersects this value.\n
 Either a date-time or an interval, open or closed. Date and time expressions adhere to RFC 3339. Open intervals are expressed using double-dots.""",  # noqa: E501
-        json_schema_extra={
-            "examples": {
-                "datetime": {"value": "2018-02-12T23:20:50Z"},
-                "closed-interval": {"value": "2018-02-12T00:00:00Z/2018-03-18T12:31:12Z"},
-                "open-interval-from": {"value": "2018-02-12T00:00:00Z/.."},
-                "open-interval-to": {"value": "../2018-03-18T12:31:12Z"},
-            },
+        openapi_examples={
+            "user-provided": {"value": None},
+            "datetime": {"value": "2018-02-12T23:20:50Z"},
+            "closed-interval": {"value": "2018-02-12T00:00:00Z/2018-03-18T12:31:12Z"},
+            "open-interval-from": {"value": "2018-02-12T00:00:00Z/.."},
+            "open-interval-to": {"value": "../2018-03-18T12:31:12Z"},
         },
     )
     limit: Optional[Limit] = Field(